+++ system/basic/libc/libewoksys/ewoksys/src/proc.c
@@ -309,7 +309,7 @@ inline uint32_t proc_get_uuid(int32_t pid) {
     return _vsyscall_info->proc_info[pid].uuid;
 }
 
-inline void* proc_malloc_expand(int32_t size) {
+inline void* proc_malloc_expand(int64_t size) {
     //klog("proc_malloc_expand %d, pid: %d\n", size, getpid());
     return (void*)syscall1(SYS_MALLOC_EXPAND, (ewokos_addr_t)size);
 }
